requirements.txt
# ── Core Framework ──────────────────────────────────────────────────────────
fastapi>=0.108.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0,<3
pydantic-settings>=2.2.0

# ── Async / Task Queue ──────────────────────────────────────────────────────
celery>=5.3.0
redis>=5.0.0
websockets>=12.0

# ── Deep Learning ───────────────────────────────────────────────────────────
# For a one-step install, use:  ./scripts/install_deps.sh
# (Installs PyTorch first, then torch-scatter/torch-sparse, then the rest.)
#
# Manual install:
#   1. pip install torch torchvision torchaudio  [or --index-url .../cu118 for GPU]
#   2. pip install torch-scatter torch-sparse -f https://data.pyg.org/whl/torch-$(python -c "import torch; print(torch.__version__.split('+')[0])")+cpu.html
#   3. pip install torch-geometric
#   4. pip install -r requirements-core.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
torch-geometric>=2.4.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0

# ── Graph / Network Analysis ────────────────────────────────────────────────
networkx>=3.2
python-louvain>=0.16
pyvis>=0.3.0

# ── Traditional ML ──────────────────────────────────────────────────────────
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.43.0
imbalanced-learn>=0.11.0

# ── Data Processing ─────────────────────────────────────────────────────────
pandas>=2.0.0
polars>=0.19.0
numpy>=1.24.0,<2
pyarrow>=14.0.0
scipy>=1.11.0

# ── Blockchain Interaction ──────────────────────────────────────────────────
web3>=6.0.0
requests>=2.28.0

# ── Database Connectors ─────────────────────────────────────────────────────
psycopg2-binary>=2.9.0
neo4j>=5.14.0
motor>=3.3.0

# ── Kafka Streaming ─────────────────────────────────────────────────────────
kafka-python>=2.0.2

# ── Monitoring ──────────────────────────────────────────────────────────────
prometheus-client>=0.19.0

# ── Testing ─────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
pytest-mock>=3.12.0

# ── Visualization (evaluation graphs) ────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.13.0

# ── Utilities ───────────────────────────────────────────────────────────────
tqdm>=4.65.0
loguru>=0.7.0
python-multipart>=0.0.9
PyJWT>=2.8.0
slowapi>=0.1.8
